refactor(config): extract global shortcut keys and mapping helpers
